tabs fix

+++ projects/timer/timer.c
@@ -33,22 +33,22 @@ int main(void);
 *************************************************/
 void TIM2_IRQHandler(void)
 {
-    static uint8_t i = 1;
+    static uint32_t i = 1;
 
-	// clear interrupt status
-	if (TIM2->DIER & 0x01) {
-		if (TIM2->SR & 0x01) {
-			TIM2->SR &= ~(1U << 0);
-		}
-	}
+    // clear interrupt status
+    if (TIM2->DIER & 0x01) {
+        if (TIM2->SR & 0x01) {
+            TIM2->SR &= ~(1U << 0);
+        }
+    }
 
     GPIOD->ODR = (i << 12);
 
     if (i == 0x08) {
-    	i = 1;
+        i = 1;
     }
     else {
-    	i = (i << 1);
+        i = (i << 1);
     }
 }
